requirements.txt
absl-py==2.1.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
boto3==1.34.56
botocore==1.34.56
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cinemagoer==2023.5.1
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
comm==0.2.1
comtypes==1.3.1
contourpy==1.2.0
cryptography==42.0.5
cv==1.0.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.2
django-allauth==65.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
docutils==0.20.1
drf-yasg==1.21.7
dynaconf==3.2.4
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.0
gspread==6.1.2
h5py==3.10.0
httplib2==0.22.0
hvac==2.1.0
idna==3.6
IMDbPY==2022.7.9
inflection==0.5.1
ipykernel==6.29.0
ipython==8.12.3
itsdangerous==2.1.2
jaraco.context==4.3.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keras==2.15.0
keyboard==0.13.5
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
libclang==16.0.6
loguru==0.7.2
lxml==5.1.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.0
scipy==1.12.0
seaborn==0.13.2
shortuuid==1.0.12
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tinycss2==1.3.0
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.25
uritemplate==4.1.1
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.6.0
wikipedia==1.4.0
win32-setctime==1.1.0
wolframalpha==5.0.0
wrapt==1.14.1
xmltodict==0.13.0
yarg==0.1.9
